Improve many of the things
This is an omnibus commit, so a lot going on. Probably the most
significant change is refactoring the Result type to be a straight
slice of ResultField with no separate Ptr field. Since we weren't
using this exported Ptr field for much, it just seems overcomplicated
to me to expose it. We'll still use it internally when deduping
partial results.