Finish basic happy path query test scenarios
There are still a lot of TODO scenarios, but the ones remaining are
the more advanced cases, e.g. where there we need to poll the chunk
more than once, or where there's multiple chunks, plus preview mode.

As part of this commit, I also refactored the non-Unmarshal related
error wrapping code in `errors.go`. This was kind of a last minute
panic thing, but it occurred to me that customers might want to be
able to differentiate "job status" type errors like "Failed" or
"Cancelled" from other fatals like "internal server error", in order
to give their customers better error messaging.

Things are in a pretty good state now. Tests passing on Linux and
Windows.