Simplify and correct mgr concurrency, eliminate deadlocks
This commit eliminates a number (hopefully all) of deadlock issues that
remained even after some previous concurrency-related refactorings. As
well, it aligns the concurrency implementation better with the Go
philosophy of concurrency. ("Do not communicate by sharing memory,
instead share memory by communicating.")

The key change in this commit is to reduce the scope of the mutex in
the mgr structure so that it now only controls access to the stats
data structure. Other mgr state previously guarded by the mutex is
now exclusively owned by the mgr `loop()` goroutine so it is no longer
shared memory and does not need a mutex.

The micro changes to make this possible are:

- Get rid of the `closed` state flag, remove buffering from the `close`
  channel, and exclusively use the `close` channel to communicate
  closed state for the mgr. Since the channel is unbuffered and only
  received from once, only one goroutine can successfully send to the
  channel. Any other goroutine attempting to send will eventually panic
  because the channel has been closed, and recovering from the panic
  allows the goroutine to return `ErrClosed`.

- Convert the previously "void" `query` channel into a channel
  containing `*streams`. This allows goroutines calling `Query` to send
  the newly-constructed query stream to the mgr `loop()` for processing
  in a lock-free manner.

Testing:

```
$ go test -count=1000 -timeout=15m
PASS
ok      github.com/gogama/incite        565.696s

$ go version
go version go1.15.6 windows/amd64
```