Ensure mgr loop shutdown closes all open streams
This commit fixes a subtle bug arising because streams can be removed
from mgr.pq yet still have in-flight chunks. The bug was that when the
mgr is closed, the loop shutdown cancels the in-flight chunks but
doesn't ensure that the stream corresponding to the chunk is marked as
closed.