Remove Hint from QuerySpec as it overcomplicates with unproven benefit
I'm doubtful the hinting, if done "correctly" by the user of the
library, would make much difference in the way of performance; and it
puts an onus on the user to optimize prematurely, or try to anyway.

Long story short, this commit simplifies the public interface and makes
the the Incite library easier to use.