Support user-specified friendly name for QueryManager in logs
This allows the user to get logs with higher information content.

The problem we're trying to solve is that if the user is fanning out
many queries across many AWS accounts in many regions, all
simultaneously, he is going to need to run multiple QueryManagers
simultaneously, since each QM is basically tied to a single set of
AWS credentials and a single AWS account and region.

When getting logs from multiple QMs in parallel, the previous
behavior of just logging the QM's address for disambiguation
doesn't do a lot to help with debugging. With this new user-
specified name, the user can give the QM a meaningful name, like
"111111111111:us-west-2:foo", which makes it much easier to
meaningfully correlate issues in the logs.