Retry failed queries [issue#5]
https://github.com/gogama/incite/issues/5

To make this work, I introduced a new queue of chunks called `ready`
living within the `mgr` structure.

The old algorithm for starting a chunk was to take the highest priority
stream off the front of `pq` and create the next chunk from the stream,
send the StartQuery request to CWL Insights, and push the new chunk onto
the back of the `chunk` polling queue (ring).

The new algorithm for starting a chunk is to take the first available
chunk from the `ready` queue (ring), send the StartQuery request to CWL
Insights, and push the chunk to the back of the `chunk` polling queue
(ring). Only if the `ready` queue is empty does a new chunk get created
from the highest priority stream in `pq`.

The introduction of the `ready` queue allows us to retry failed chunks
fairly easily, by simply putting the failed chunks at the back of the
ready queue so they will be restarted.