Add and document API changes for splitting and progress (issues #3 + #1)
https://github.com/gogama/incite/issues/3
https://github.com/gogama/incite/issues/1

I probably shouldn't have coupled these together in a single commit, but
for some reason I see them as releated.

Chunk splitting support is added in `QuerySpec` as the `Split` and
`ReSplit` fields. Validation is done and tested, but no implementation
is done.

Progress transparency is added in the `Stats` structure as the four
new fields prefixed with `Range`. I thought about various ways of
indicating progress but since conceptually each query represents a scan
across a continguous time range, using `time.Duration` seemed to make
the most sense in the end. Also because it makes just as much sense
for `QueryManager` as for `Stream`.