Improve dynamic splitting API (issue #3)
https://github.com/gogama/incite/issues/3

The previous edition was too complicated, and consequently not very
user-friendly. If you're trying to query some logs, the last thing you
want to do is have to configure two separate parameters based on powers
of two and summing. The last straw for me was I couldn't imagine how one
could expose both of these parameters in a comprehensible way to users
of a command line interface.

Apart from this API rewrite, this commit also fixes an issue with the
Chunk validation where it failed to prevent sub-second granularities in
chunks, which could theoretically cause odd bugs.