Refactor concurrency model for performance and comprehension (issue #9)
https://github.com/gogama/incite/issues/9

This commit gets us most of the way there, although there are a few test
cases that continue to fail sporadically, some missing unit tests, and
some TODO items still needing to be handled.

In a nutshell, I have divided the `mgr` logic that previously ran within
a single goroutine (from `mgr.loop()`) into four pieces. Orchestration
logic still runs from the `mgr` goroutine, but there are now three other
"worker" goroutines that make the web service calls to the CloudWatch
Logs Insights API: starter, poller, and stopper.

The mgr goroutine now sends chunks to the starter for starting, and the
starter sends them back to the mgr loop once a start result is obtained.
The mgr goroutine sends started chunks down to the poller for polling,
and the poller sends them back to the mgr loop once a final polling
result is obtained. The mgr goroutine sends running chunks to the
stopper if they need to be aborted for some reason.

As a result of this refactor, a single mgr is able to use all available
CloudWatch Logs API capacity for starting and polling simultaneously.
Previously, it had been prioritizing starting chunks over polling
them, which meant that even if many chunks queries finished running
simultaneously, as soon as the mgr loop noticed that one was finished,
it would prioritize starting another one, which would tend to delay
service to other chunks that were finished running...

The success metric I had set in advance for this refactor was any
material reduction in the time to run one iteration of the unit tests.
Right before the refactor the tests had been taking 4.661 seconds on my
laptop, and this commit brings the single iteration time down to about
1.712 seconds, a 63% reduction in runtime. So despite a few loose ends
to tie up, this refactor has been a success.