Lift chunk split logic up from poller into mgr
The purpose of moving the logic is to make it easier to reason about
possible deadlock.

When splitting lives in the poller, every chunk sent to the poller
can result in `1+splitBits` chunks being sent back on the update
channel. In a degenerate case with tonnes of splitting, this could
result in number of chunks in the update channel buffer essentially
growing without bound, which can eventually result in deadlock if
`poller` is blocked trying to put into the update channel while
`mgr` is blocked trying to put into the poller channel. By pulling
the splitting up into `mgr` we avoid this scenario.