Fix bug where poller could cause chunks to be lost (issue #9)
https://github.com/gogama/incite/issues/9

This bug was introduced in the concurrency model refactor. The problem
arises when poller gets an in-progress status back from Insights for
the same query ID more times than the poller's retry count. This isn't
an error, and doesn't result in an error being attached to the chunk,
but it causes the poller to send the chunk back to the mgr without it
being in in the complete state. The mgr treats the lack of complete
state as an error condition and tries to kill the stream. But, since
there is no error, the stream doesn't get killed. Moreover, if this
is the only chunk in the stream, the stream also doesn't get unblocked
and so any calls to Read() will wait forever.

This commit fixes the problem. However, the hunt continues for at
least one other issue that can freeze the tests at high reps, plus
an issue that causes leaking goroutines...