Fix goroutine leak in a test case
The test case "Query Starts But GetQueryResults Output Corrupt" under
TestQueryManager_Query was failing to close the query manager. Since
there are 5x2 = 10 sub-cases under this test, this led to leaking 40
goroutines each iteration - 10 of each kind of worker plus 10 instances
of the mgr loop.

This commit fixes that leak and rearranges the code in a handful of
other places to bring the cleanup closer to the query manager
instantiation. It also adds a rudimentary sniff test for goroutine
leaks.