Fix occasional deadlock during shutdown (issue #9)
https://github.com/gogama/incite/issues/9

This is related to, but slightly different from, the bug fixed in commit
f4fab079, and is another bug introduced in the concurrency refactor
process.

If, during the mgr/worker shutdown process, a worker notices the closure
of the close channel before the mgr, and the worker is able to complete
its shutdown function while the mgr loop function is still running, then
the chunks released by the worker shutdown process often will not
contain an error. This means that when they get back to the mgr, there
is no chunk error to set on the stream and in some cases this can result
in th stream not being killed properly, meaning that goroutines blocked
on the condition wait in stream.Read stay there forever.

This commit fixes the problem by having the worker shutdown assign all
released chunks the error ErrClosing.