Fix bug unit test due to chunk alignment change (issue #11)
https://github.com/gogama/incite/issues/11

The previous commit caused a test failure that only manifests sometimes
because TestStream_Close has an inherent race condition. Basically if
the starter managed to call StartQueryWithContext before it detects
shutdown, then the test would fail with the below message:

    panic: Fail in goroutine after TestStream_Close has completed

But if shutdown happened first, the test would pass. This commit fixes
the problem.