Finish bug fix RE: failure to respect parallelism limit
The parallism limit (`Config.Parallel`, `mgr.Parallel`) is intended to
ensure the query manager doesn't exceed the CloudWatch Logs Insights
service quota limit on number of running queries.

The mgr's best estimate of the (upper bound on the) number of running
queries is the sum of the number of queries it is trying to start, the
number it has started and is polling, and the number that it has started
but is now trying to stop for some reason. So in order to ensure the mgr
never has more balls in the air than the `Config.Parallel` limit, it
needs to pause its starting activities as soon as that sum adds up to
the parallelism limit.

To put it another way, the earlier commit 8abbdc5a was a partial fix to
the problem but not a complete fix. This commit should complete the fix.