Denote chunk time range in epoch milliseconds (issue #16)
https://github.com/gogama/incite/issues/16

This change should fix a bug in which 999 milliseconds at the end of the
chunk boundary were being lost.

This bug originated in inaccurate CWL documentation and a buggy
implementation of their inaccurate documentation on the server side, but
as I haven't been successful in prompting them to fix it, this should
eliminate the problem on Incite's end.

Two further notes:

1. I'll be doing integration testing against a live API and posting the
   results to issue #16 before closing the issue. This is not done as of
   the time of this commit.
2. I will make a subsequent small improvement to change the sub-second
   granularity limit to sub-millisecond granularity. This is possible
   since we are now expressing the time range at millisecond
   granularity.