Dynamically adapt requests per second to noisy neighbors (issue#23)
https://github.com/gogama/incite/issues/23

This change modifies the three worker goroutines (starter, poller,
stopper) so that the maximum number of requests per second they try
to send to the CloudWatch Logs Insights web service adapts downward
when a request gets throttled by the web service, and upward when it
does not.

The hope is that with this change, QueryManager.Query() will become more
resilient. If a given chunk collects more than about 10 (e.g.
`maxTempPollingErrs`) temporary errors, the owning stream errors out,
which can be very annoying if it has been running for a long time and
doing a lot of work. This change does not change that fundamental
behavior, but it does make it much less likely that the chunk will
collect such a large number of temporary errors, by ensuring the
worker backs off if it is getting throttled.

In future, we can twiddle with the specific constant values both in
terms of `maxTempPollingErrs` and the like and the adaptive RPS
configuration, but hopefully the core idea here will stand the test
of time.

NOTE: This commit does not complete issue #23. To complete the issue,
we also need to dynamically adapt `Parallel` so that `mgr` also backs
off when it hits CWL Insights query concurrency limits. Concurrency
limits are the main reason why chunks collect more than
`maxTempStartingErrs` in the starter worker.