Fix end-before-start bug  (issue#25)
https://github.com/gogama/incite/issues/25

As suggested, the fix was to separate `mgr.n` into two fields to
reflect the two competing jobs it was trying to perform. (Tracking
the number of generation 0 chunks on one hand, for the purpose of
determing whether to remove a stream from `mgr.pq`; and tracking
the total number of known chunks for the stream, for the purpose of
knowing whether the chunk is EOF.)

In implementing this commit, I realized that `mgr.n` was incorrectly
listed in the "immutable" section of `type mgr struct`, so that has
also been fixed. This incorrect placement would have occurred
stealthily during implementation of dynamic chunk splitting.