Dynamically adapt query parallelism to noisy neighbors (issue#23)
https://github.com/gogama/incite/issues/23

This change modifies the mgr to dynamically change the amount of
attempted parallelism based on whether it is able to start chunks
without receiving a service quota limited exceeded error from the
CloudWatch Logs Insights Service's `StartQuery` API operation.

As with the earlier commit `8416966a`, the hope is that this commit
will make `QueryManager.Query()` more resilient. Previously, it would
be very annoying if a chunk collected more than 10 (i.e.
`maxTempStartingErrors`) temporary errors when the starter was trying
to start it. This could occur if other users were heavily querying the
CloudWatch Logs Insights service in the same AWS region, resulting in
many "LimitExceededException" errors due to exceeding the service's
query concurrency quota limit. When this would happen, it would kill
the entire stream, resulting in the end user's entire `Query()`
operation failing.

With this change, a mgr's effective parallelism can temporarily go
below the value given in `Config.Parallel`, but will increase back
up as chunks are successfully started. It will never go above the
value given in `Config.Parallel`. Temporary reductions are made in
response to "LimitExceededException" errors and should cause the mgr
to adjust its attempted chunk parallelism to the amount of free
concurrency that is actually available, accounting for other "noisy
neighbors" who are simultaneously using the Insights service,
including other instances of the Incite QueryManager. This should lead
to fewer "LimitExceededException" errors being received, reducing the
probability of a `Query()` failure.